chore(tests): split up e2e tests - tf amendments
