+++ .github/workflows/test.yml
@@ -582,9 +582,7 @@ jobs:
             'case-amendments/facility-end-date/feature-flag-enabled/*.spec.js',
             'case-amendments/in-progress/*.spec.js',
             'case-amendments/request-when-previously-submitted/*.spec.js',
-            'case-amendments/show-amendment-button/AIN/*.spec.js',
-            'case-amendments/show-amendment-button/MIA/*.spec.js',
-            'case-amendments/show-amendment-button/MIN/*.spec.js',
+            'case-amendments/show-amendment-button/*.spec.js',
             'case-amendments/tasks/*.spec.js',
             'case-amendments/underwriting-page-unassigned/*.spec.js',
             'case-amendments/user-flow/*.spec.js',
