--- e2e-tests/tfm/cypress/e2e/journeys/case-amendments/show-amendment-button/MIA/show-amendment-button-MIA.spec.js
@@ -1,50 +1,0 @@
-import relative from '../../../../relativeURL';
-import facilityPage from '../../../../pages/facilityPage';
-import amendmentsPage from '../../../../pages/amendments/amendmentsPage';
-import MOCK_DEAL_AIN from '../../../../../fixtures/deal-AIN';
-import MOCK_DEAL_MIA from '../../../../../fixtures/deal-MIA';
-import { T1_USER_1, PIM_USER_1, BANK1_MAKER1, ADMIN } from '../../../../../../../e2e-fixtures';
-
-context('Amendments page - show amendment button - MIA', () => {
-  let dealId;
-  const dealFacilities = [];
-
-  before(() => {
-    cy.insertOneDeal(MOCK_DEAL_MIA, BANK1_MAKER1).then((insertedDeal) => {
-      dealId = insertedDeal._id;
-
-      const { dealType, mockFacilities } = MOCK_DEAL_AIN;
-
-      cy.createFacilities(dealId, mockFacilities, BANK1_MAKER1).then((createdFacilities) => {
-        dealFacilities.push(...createdFacilities);
-      });
-
-      cy.submitDeal(dealId, dealType, PIM_USER_1);
-    });
-  });
-
-  after(() => {
-    cy.deleteDeals(dealId, ADMIN);
-    dealFacilities.forEach((facility) => {
-      cy.deleteFacility(facility._id, BANK1_MAKER1);
-    });
-  });
-
-  it('should NOT render an `add amendment` button when PIM user', () => {
-    cy.login(PIM_USER_1);
-    const facilityId = dealFacilities[0]._id;
-    cy.visit(relative(`/case/${dealId}/facility/${facilityId}`));
-
-    facilityPage.facilityTabAmendments().click();
-    amendmentsPage.addAmendmentButton().should('not.exist');
-  });
-
-  it('should NOT render an `add amendment` button when NOT a PIM user', () => {
-    cy.login(T1_USER_1);
-    const facilityId = dealFacilities[0]._id;
-    cy.visit(relative(`/case/${dealId}/facility/${facilityId}`));
-
-    facilityPage.facilityTabAmendments().click();
-    amendmentsPage.addAmendmentButton().should('not.exist');
-  });
-});
